backend/requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# Cache & Queue
redis==5.0.1
celery==5.3.6

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.1.2
cryptography==42.0.0

# HTTP Client
requests==2.31.0
httpx==0.26.0

# XML Processing (for eBay Trading API)
lxml==5.1.0

# Rate Limiting
slowapi==0.1.9

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Code Quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Task Scheduling
APScheduler==3.10.4
